requirements.txt
fpdf==1.7.2
RPi.GPIO==0.7.1
pygame==2.1.2
Pillow==9.3.0
